Revert "43.login to profile functionality"
