+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".ViewProfileActivity"></activity>
-        <activity android:name=".CreateAccountActivity"></activity>
         <activity
             android:name=".ViewVibeActivity"
             android:label="@string/view_vybe_name"
