+++ app/src/main/java/com/example/vybe/LogInActivity.java
@@ -1,0 +1,43 @@
+package com.example.vybe;
+
+import android.os.Bundle;
+
+import android.util.Log;
+import android.widget.EditText;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+/**
+ * This activity gets the username and password and validates said fields
+ * with Firestore. Still WIP
+ */
+public class LogInActivity extends AppCompatActivity {
+
+    private static final String TAG = "LoginActivity";
+
+    private FirebaseAuth mAuth;
+
+    private EditText username_input, password_input;
+
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        // Check if user is signed in (non-null) and update UI accordingly.
+        FirebaseUser currentUser = mAuth.getCurrentUser();
+        //updateUI(currentUser);
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_sign_in);
+        username_input = findViewById(R.id.username_field);
+        password_input = findViewById(R.id.password_field);
+
+        mAuth = FirebaseAuth.getInstance();
+    }
+}
