+++ app/src/main/res/layout/activity_main.xml
@@ -15,8 +15,8 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:layout_weight="0"
-        android:padding="5dp"
-        android:src="@drawable/ic_title_vibe" />
+        android:padding="10dp"
+        android:src="@drawable/ic_title_vibe" />t
 
     <TextView
         android:id="@+id/app_title"
@@ -26,22 +26,8 @@
         android:textColor="@color/White"
         android:textStyle="bold"
         android:layout_gravity="center_horizontal"
-        android:padding="25dp"
+        android:padding="35dp"
         android:textSize="48sp" />
-
-    <EditText
-        android:id="@+id/email_edit_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="Email"
-        android:inputType="textEmailAddress" />
-
-    <EditText
-        android:id="@+id/password_edit_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="Password"
-        android:inputType="textPassword" />
 
     <Button
         android:id="@+id/login_button"
