+++ docs/javadocs/com/example/vybe/LogInActivity.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.example.vybe.CreateAccountActivity</li>
+<li>com.example.vybe.LogInActivity</li>
 </ul>
 </li>
 </ul>
